Enhance TurnstileCaptcha component by implementing compact size for mobile layout, improving debug information for mobile rendering, and refining margin styles for better UI consistency. Maintain TypeScript standards and improve overall code structure.
